AnnaChive---Open-Source-Resource-Downloader-1.0.0/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "annchive"
version = "0.1.0"
description = "Local CLI tool for downloading open-access resources with Tor anonymity"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Akarsh", email = "akarsh@example.com"}
]
keywords = ["cli", "open-access", "annas-archive", "tor", "download"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: Proxy",
]
dependencies = [
    "click>=8.1.0",
    "httpx>=0.27.0",
    "stem>=1.8.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "cryptography>=44.0.0",
    "keyring>=25.0.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    "pycryptodome>=3.20.0",  # For torrent handling
]

[project.optional-dependencies]
torrent = [
    "libtorrent>=2.0.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "mypy>=1.8.0",
]

[project.scripts]
annchive = "annchive.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["annchive*"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true